% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_loc.R
\name{count_loc}
\alias{count_loc}
\title{Count lines of code}
\usage{
count_loc(
  paths = ".",
  excluded = NULL,
  include_hidden = FALSE,
  languages = "R"
)
}
\arguments{
\item{paths}{Paths to include.}

\item{excluded}{Paths to exclude.}

\item{include_hidden}{Count hidden files and directories? Default is \code{FALSE}.}

\item{languages}{Languages to consider. See the list here: \url{https:#github.com/XAMPPRocky/tokei?tab=readme-ov-file#supported-languages}.}
}
\value{
A data.frame with 4 columns: number of blank lines, comments, code,
and whether the count is accurate.
}
\description{
Count lines of code
}
\examples{
tmp <- tempfile(fileext = ".R")
cat("
library(loc)
# This is a function
foo <- function(x) {
  x + 1
}

", file = tmp)
count_loc(tmp)
}
