% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{loc-package}
\alias{loc}
\alias{loc-package}
\title{loc: Count Lines of Code}
\description{
Use the 'Tokei' Rust crate to count lines of code in files and folders. It supports dozens of different languages, and also provides statistics on blank lines and comments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/etiennebacher/loc}
  \item Report bugs at \url{https://github.com/etiennebacher/loc/issues}
}

}
\author{
\strong{Maintainer}: Etienne Bacher \email{etienne.bacher@protonmail.com} [copyright holder]

}
