% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titles.R
\name{loc_search_pages}
\alias{loc_search_pages}
\title{Search}
\usage{
loc_search_pages(q, pages = 1, sleep = get_sleep())
}
\arguments{
\item{q}{Search terms}

\item{pages}{Number of pages of results to collect.}

\item{sleep}{Number of milliseconds to sleep between \code{pages}.}
}
\description{
There are millions of digitized newspaper pages in Chronicling America. 
These pages span several decades and many U.S. states and territories. 
New batches of data come in from partner institutions throughout the year 
and are added to the site regularly.
}
\examples{
\dontrun{articles <- loc_search_pages("thomas", pages = 2)}

}
