% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titles.R
\name{loc_search_titles}
\alias{loc_search_titles}
\title{Titles}
\usage{
loc_search_titles(q, pages = 1, sleep = get_sleep())
}
\arguments{
\item{q}{Search terms}

\item{pages}{Number of pages of results to collect.}

\item{sleep}{Number of milliseconds to sleep between \code{pages}.}
}
\description{
The directory of newspaper titles contains nearly 140,000 records of 
newspapers and libraries that hold copies of these newspapers. 
The title records are based on MARC data gathered and enhanced 
as part of the NDNP program.
}
\examples{
\dontrun{articles <- loc_search_titles("michigan", pages = 2)}

}
