% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loc_url}
\alias{loc_url}
\title{Library of Congress Base URL}
\format{A vector}
\source{
\url{https://chroniclingamerica.loc.gov}
}
\usage{
loc_url
}
\description{
Library of Congress Base URL
}
\keyword{datasets}
