% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titles.R
\name{sleep}
\alias{sleep}
\alias{set_sleep}
\alias{get_sleep}
\title{Sleep}
\usage{
set_sleep(s)

get_sleep()
}
\arguments{
\item{s}{Seconds to sleep between pages.}
}
\description{
Set and get sleep time for subsequent calls.
}
